fix(ui): remove redundant tags from explore page also update folder structure
