+++ openmetadata-ui/src/main/resources/ui/src/components/Tag/Tags/tags.interface.ts
@@ -11,22 +11,18 @@
  *  limitations under the License.
  */
 
-import { EntityTags, TagOption } from 'Models';
-import { ReactNode } from 'react';
-import { TagProps } from '../tags/tags.interface';
+import { TagLabel } from '../../../generated/type/tagLabel';
 
-export type TagsContainerProps = {
-  children?: ReactNode;
-  editable?: boolean;
-  dropDownHorzPosRight?: boolean;
-  selectedTags: Array<EntityTags>;
-  tagList: Array<TagOption | string>;
-  type?: TagProps['type'];
-  showTags?: boolean;
-  showAddTagButton?: boolean;
+export type TagProps = {
   className?: string;
-  containerClass?: string;
-  buttonContainerClass?: string;
-  onSelectionChange?: (selectedTags: Array<EntityTags>) => void;
-  onCancel?: (event: React.MouseEvent<HTMLElement, MouseEvent>) => void;
+  editable?: boolean;
+  type?: 'contained' | 'outlined' | 'label' | 'border';
+  startWith?: '#' | '+ ';
+  tag: string | TagLabel;
+  isRemovable?: boolean;
+  showOnlyName?: boolean;
+  removeTag?: (
+    event: React.MouseEvent<HTMLElement, MouseEvent>,
+    removedTag: string
+  ) => void;
 };
