+++ openmetadata-ui/src/main/resources/ui/src/components/Tag/Tags/tags.styles.ts
