+++ openmetadata-ui/src/main/resources/ui/src/pages/service/index.tsx
@@ -32,7 +32,7 @@ import Loader from 'components/Loader/Loader';
 import { usePermissionProvider } from 'components/PermissionProvider/PermissionProvider';
 import { OperationPermission } from 'components/PermissionProvider/PermissionProvider.interface';
 import ServiceConnectionDetails from 'components/ServiceConnectionDetails/ServiceConnectionDetails.component';
-import TagsViewer from 'components/tags-viewer/tags-viewer';
+import TagsViewer from 'components/Tag/TagsViewer/tags-viewer';
 import { t } from 'i18next';
 import { isEmpty, isNil, isUndefined, startCase, toLower } from 'lodash';
 import { ExtraInfo, ServicesUpdateRequest, ServiceTypes } from 'Models';
